requirements.txt
Jinja2
click
prompt_toolkit
python-box
